+++ packages/input-autocomplete/src/index.ts
@@ -1,1 +1,5 @@
-export * from './Component';
+export {
+    InputAutocompleteDesktop as InputAutocomplete,
+    InputAutocompleteDesktopProps as InputAutocompleteProps,
+} from './Component.desktop';
+export * from './responsive';