+++ packages/select/src/components/checkmark/Component.tsx
@@ -1,9 +1,10 @@
 import React, { useCallback } from 'react';
 import cn from 'classnames';
 
+import { Badge } from '@alfalab/core-components-badge';
 import { Checkbox, CheckboxProps } from '@alfalab/core-components-checkbox';
-import { Badge } from '@alfalab/core-components/badge';
 import { CheckmarkCircleMIcon } from '@alfalab/icons-glyph/CheckmarkCircleMIcon';
+
 import { CheckmarkProps } from '../../typings';
 
 import styles from './index.module.css';