+++ packages/date-range-input/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @alfalab/core-components-date-range-input
 
+## 2.2.0
+
+### Minor Changes
+
+### [#334](https://github.com/core-ds/core-components/pull/334)
+
+-   В компонентах DateRangeInput, DateTimeInput и CalendarInput добавлены mobile и desktop версии компонентов. Название компонентов было изменено по схеме Component → ComponentDesktop
+-   В компонентах Calendar и InputAutocomplete добавлена responsive версия компонентов. Название компонентов было изменено по схеме Component → ComponentResponsive<br />
+-   В компонентах PickerButton и Tooltip добавлен новый пропс breakpoint. Название компонентов было изменено по схеме Component -→ ComponentResponsive для PickerButton и Component → ComponentDesktop для Tooltip<br />
+-   В компонентах Confirmation и SidePanel добавлен новый пропс breakpoint<br />
+
+### Patch Changes
+
+### [#418](https://github.com/core-ds/core-components/pull/418)
+
+-   Исправлена проблема с default-импортом в cjs форматах
+
+-   Обновлены зависимости
+    -   calendar@6.2.0
+    -   icon-button@6.0.2
+    -   input@11.1.1
+    -   popover@6.0.5
+
+## 2.1.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   calendar@6.1.15
+
 ## 2.1.2
 
 ### Patch Changes