+++ packages/tooltip/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Change Log
 
+## 6.2.0
+
+### Minor Changes
+
+### [#334](https://github.com/core-ds/core-components/pull/334)
+
+-   В компонентах DateRangeInput, DateTimeInput и CalendarInput добавлены mobile и desktop версии компонентов. Название компонентов было изменено по схеме Component → ComponentDesktop
+-   В компонентах Calendar и InputAutocomplete добавлена responsive версия компонентов. Название компонентов было изменено по схеме Component → ComponentResponsive<br />
+-   В компонентах PickerButton и Tooltip добавлен новый пропс breakpoint. Название компонентов было изменено по схеме Component -→ ComponentResponsive для PickerButton и Component → ComponentDesktop для Tooltip<br />
+-   В компонентах Confirmation и SidePanel добавлен новый пропс breakpoint<br />
+
+### Patch Changes
+
+### [#418](https://github.com/core-ds/core-components/pull/418)
+
+-   Исправлена проблема с default-импортом в cjs форматах
+
+-   Обновлены зависимости
+    -   bottom-sheet@5.1.13
+    -   button@7.0.2
+    -   popover@6.0.5
+
+## 6.1.13
+
+### Patch Changes
+
+### [#402](https://github.com/core-ds/core-components/pull/402)
+
+-   Добавлено свойство `max-width` для Tooltip.
+
 ## 6.1.12
 
 ### Patch Changes